全世界データのstateをApp.jsxに移動。それに伴い、WorldPage、Cardsコンポーネントの変更
