+++ src/App.jsx
@@ -1,9 +1,9 @@
-import './App.css'
-import TopPage from './pages/TopPage'
-import WorldPage from './pages/WorldPage';
 import {useState} from 'react';
 import {Route, Routes, BrowserRouter} from 'react-router-dom';
 import countriesJson from './countries.json';
+import TopPage from './pages/TopPage'
+import WorldPage from './pages/WorldPage';
+import './App.css'
 
 function App() {
   // Selctorコンポーネントで選ばれた国のデータを受け取るためのstate
@@ -17,16 +17,28 @@ function App() {
     totalRecovered: "",
   });
 
+  // 全世界のデータを受け取るstateを設定
+  const [allCountriesData, setAllCountriesData] = useState([]);
+
   // その国データを受け取る  
   const getCountryData = () => {
-      fetch(`https://monotein-books.vercel.app/api/corona-tracker/country/${country}`).then(res => res.json()).then(data => { setCountryData({
+    fetch(`https://monotein-books.vercel.app/api/corona-tracker/country/${country}`)
+    .then(res => res.json())
+    .then(data => { setCountryData({
         date: data[data.length -1].Date,
         newConfirmed: data[data.length -1].Confirmed - data[data.length -2].Confirmed,
         totalConfirmed: data[data.length -1].Confirmed,
         newRecovered: data[data.length -1].Recovered - data[data.length -2].Recovered,
         totalRecovered: data[data.length -1].Recovered,
-        }); 
-      });
+      }); 
+    });
+  }
+
+  // 全世界のデータを取得する
+  const getAllCountriesData = () => {
+    fetch("https://monotein-books.vercel.app/api/corona-tracker/summary")
+    .then(res => res.json())
+    .then(data => setAllCountriesData(data.Countries))
   }
 
   return (
@@ -34,11 +46,10 @@ function App() {
       <Routes>
         <Route path="/" element={<TopPage countriesJson={countriesJson} setCountry={setCountry}  getCountryData={getCountryData} countryData={countryData} />} />
         <Route path="/world" element={
-          <WorldPage />
+          <WorldPage allCountriesData={allCountriesData} getAllCountriesData={getAllCountriesData}/>
         } />
       </Routes>
     </BrowserRouter>
-
   )
 }
